+++ OrangeDesignSystemDemo/OrangeDesignSystemDemo/Screens/Components/Pages/Cards/CardVerticalHeaderFirstVariant.swift
@@ -32,13 +32,16 @@ class CardVerticalHeaderFirstVariantModel: ObservableObject {
 
     @Published var showThumbnail: Bool
     @Published var showSubtitle: Bool
-    @Published var showSupportingText: Bool
+    @Published var showText: Bool
     @Published var buttonCount: Int
     @Published var showAlert: Bool
-    var alertText: String = ""
-
-    private let buttonsText = ["Button 1", "Button 2"]
     
+    var alertText: String = ""
+    let buttonsText = ["Button 1", "Button 2"]
+    private var recipe: Recipe {
+        RecipeBook.shared.recipes[0]
+    }
+
     // =================
     // MARK: Initializer
     // =================
@@ -46,34 +49,41 @@ class CardVerticalHeaderFirstVariantModel: ObservableObject {
     init() {
         showThumbnail = true
         showSubtitle = true
-        showSupportingText = true
+        showText = true
         buttonCount = 2
         showAlert = false
     }
+        
+    // ==================
+    // MARK: Card Content
+    // ==================
     
-    // =============
-    // MARK: Helpers
-    // =============
-
-    var cardModel: ODSCardVerticalHeaderFirstModel {
-        ODSCardVerticalHeaderFirstModel(
-            title: cardExampleTitle,
-            subtitle: showSubtitle ? cardExampleSubtitle : nil,
-            thumbnail: showThumbnail ? Image("ods_empty", bundle: Bundle.ods) : nil,
-            imageSource: cardExampleImage,
-            supportingText: showSupportingText ? cardExampleSupportingText : nil)
+    var title: Text {
+        Text(recipe.title)
     }
     
-    var button1Text: String? {
-        buttonCount >= 1 ? buttonsText[0] : nil
+    var subtitle: Text? {
+        showSubtitle ? Text(recipe.subtitle) : nil
     }
-
-    var button2Text: String? {
-        buttonCount >= 2 ? buttonsText[1] : nil
+    
+    var thumbnail: Image? {
+        showThumbnail ? Image("ods_empty", bundle: Bundle.ods) : nil
     }
-
-    var numberOfButtons: Int {
-        buttonsText.count
+    
+    var imageSource: ODSImage.Source {
+        .asyncImage(recipe.url, Image("ods_empty", bundle: Bundle.ods))
+    }
+    
+    var text: Text? {
+        showText ? Text(recipe.description) : nil
+    }
+    
+    var firstButtonText: String {
+        buttonsText[0]
+    }
+    
+    var secondButtonText: String  {
+        buttonsText[1]
     }
 
     func displayAlert(text: String) {
@@ -83,44 +93,75 @@ class CardVerticalHeaderFirstVariantModel: ObservableObject {
 }
 
 struct CardVerticalHeaderFirstVariant: View {
-
+    
     // =======================
     // MARK: Stored properties
     // =======================
-
+    
     @ObservedObject var model: CardVerticalHeaderFirstVariantModel
-
+    
     // ==========
     // MARK: Body
     // ==========
-
+    
     var body: some View {
         CustomizableVariant {
             ScrollView {
-                ODSCardVerticalHeaderFirst(model: model.cardModel) {
-                    if let text = model.button1Text {
-                        ODSButton(text: Text(text), emphasis: .low) {
-                            model.displayAlert(text: "Button 1 clicked")
-                        }
+                card
+                    .padding(.horizontal, ODSSpacing.m)
+                    .padding(.top, ODSSpacing.m)
+                    .onTapGesture {
+                        model.displayAlert(text: "Card container clicked")
                     }
-                } buttonContent2: {
-                    if let text = model.button2Text {
-                        ODSButton(text: Text(text), emphasis: .low) {
-                            model.displayAlert(text: "Button 2 clicked")
-                        }
-                    }
-                }
-                .padding(.horizontal, ODSSpacing.m)
-                .padding(.top, ODSSpacing.m)
-                .onTapGesture {
-                    model.displayAlert(text: "Card container clicked")
-                }
             }
             .alert(model.alertText, isPresented: $model.showAlert) {
                 Button("close", role: .cancel) {}
             }
         } options: {
             CardVerticalHeaderFirstVariantOptions(model: model)
+        }
+    }
+    
+    @ViewBuilder
+    private var card: some View {
+        switch model.buttonCount {
+        case 0:
+            ODSCardVerticalHeaderFirst(title: model.title,
+                                       imageSource: model.imageSource,
+                                       subtitle: model.subtitle,
+                                       thumbnail: model.thumbnail,
+                                       text: model.text)
+        case 1:
+            ODSCardVerticalHeaderFirst(
+                title: model.title,
+                imageSource: model.imageSource,
+                subtitle: model.subtitle,
+                thumbnail: model.thumbnail,
+                text: model.text
+            ) {
+                Button(model.firstButtonText) {
+                    model.displayAlert(text: "\(model.firstButtonText) clicked")
+                }
+            }
+        case 2:
+            ODSCardVerticalHeaderFirst(
+                title: model.title,
+                imageSource: model.imageSource,
+                subtitle: model.subtitle,
+                thumbnail: model.thumbnail,
+                text: model.text
+            ) {
+                Button(model.firstButtonText) {
+                    model.displayAlert(text: "\(model.firstButtonText) clicked")
+                }
+            } secondButton: {
+                Button(model.secondButtonText) {
+                    model.displayAlert(text: "\(model.secondButtonText) clicked")
+                }
+            }
+            
+        default:
+            EmptyView()
         }
     }
 }
@@ -141,11 +182,11 @@ private struct CardVerticalHeaderFirstVariantOptions: View {
         VStack(spacing: ODSSpacing.m) {
             Toggle("Thumbnail", isOn: $model.showThumbnail)
             Toggle("Subtitle", isOn: $model.showSubtitle)
-            Toggle("Text", isOn: $model.showSupportingText)
+            Toggle("Text", isOn: $model.showText)
             
             Stepper("Number of buttons: \(model.buttonCount)",
                     value: $model.buttonCount,
-                    in: 0 ... model.numberOfButtons)
+                    in: 0 ... model.buttonsText.count)
         }
         .odsFont(.bodyRegular)
         .padding(.vertical, ODSSpacing.m)