+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased](https://github.com/Orange-OpenSource/ods-ios/compare/0.13.1...qualif)
+
+
 ## [0.13.1](https://github.com/Orange-OpenSource/ods-ios/compare/0.13.1...0.12.0) - 2023-09-13
 
 - [DemoApp] Update AppNews for release