+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/Orange-OpenSource/ods-ios/compare/0.13.1...qualif)
 
+- [DemoApp/SDK] Icon Button Disables not setup ([#426](https://github.com/Orange-OpenSource/ods-ios/issues/426)) 
 - [Doc] Add documentation for release process ([#451](https://github.com/Orange-OpenSource/ods-ios/issues/451))
 - [DemoApp/SDK] About module: Review The ShareTheApp content because url is presented twice ([#435](https://github.com/Orange-OpenSource/ods-ios/issues/435))