+++ CHANGELOG.md
@@ -5,8 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/Orange-OpenSource/ods-ios/compare/0.13.1...qualif)
 
-- [SDK] Accessibility Voice over - Application name is a header in the about screen F>>>> 
-(Bug [#468](https://github.com/Orange-OpenSource/ods-ios/issues/468))
+- [SDK] Update banner api to use SwiftUI elements [#473](https://github.com/Orange-OpenSource/ods-ios/issues/473))
+- [SDK] Accessibility Voice over - Application name is a header in the about screen  (Bug [#468](https://github.com/Orange-OpenSource/ods-ios/issues/468))
 - [DemoApp/SDK] Update the button emphasis scale naming ([#464](https://github.com/Orange-OpenSource/ods-ios/issues/464))
 - [DemoApp] Update configuration to display by default buttons in variable width and without icon  ([#459](https://github.com/Orange-OpenSource/ods-ios/issues/459))
 - [DemoApp/SDK] Icon Button Disables not setup ([#426](https://github.com/Orange-OpenSource/ods-ios/issues/426))